Falcon's Shinesplosion moved to Debug Menu option
Achilles originally created the alternate SDR Falcon, with the 'Shinesplosion' down-B move, simply because the SDR Falcon wasn't really much different from vanilla Falcon, and he thought it'd be a cool and fun use of the extra slot. The same goes for the reversed shines for SDR Fox & Falco. These were never meant to be part of the SDR balancing mod. This was of course cause for a lot of confusion with SDR and Falcon (and problems with SDR tournaments), and it therefore makes a lot more sense for it to be its own option in the Debug Menu instead.